pyproject.toml
[project]
name = "learning-django"
version = "0.1.0"
description = "Django_for_profesional"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "crispy-bootstrap5>=2025.6",
    "django>=6.0.1",
    "django-allauth>=65.13.1",
    "django-allauth-ui>=1.8.1",
    "django-crispy-forms>=2.5",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
]
